Dispatch messages directly in CallbackResponseHandler
